Module pallet_proxy_genesis_companion::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

Pallet for configuring proxy at genesis

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Traits

Config

This pallet requires

Type Definitions

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.